pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xmoltoppm"
version = "1.0.0"
description = "XYZ trajectory to PPM/PNG molecular visualization (Python port of xmoltoppm)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
]

[project.optional-dependencies]
png = ["Pillow>=9.0"]

[project.scripts]
xmoltoppm = "xmoltoppm.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["xmoltoppm*"]
